apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Multi-Agent IAM Dashboard Backend API"
readme = "README.md"
authors = [
    { name = "Multi-Agent IAM Dashboard Team", email = "team@example.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.4",
    "celery>=5.5.0",
    "email-validator>=2.1.1",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.4.0",
    "pyjwt>=2.8.0",
    "python-multipart>=0.0.9",
    "redis>=5.0.0",
    "sqlmodel>=0.0.21",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "httpx>=0.28.1",
    "mypy>=1.17.1",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.7",
]

# Ruff configuration
[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "TID", # flake8-tidy-imports
    "Q",   # flake8-quotes
    "FBT", # flake8-boolean-trap
    "PL",  # pylint
    "PT",  # flake8-pytest-style
]
ignore = [
    "E501",   # line too long, handled by formatter
    "B008",   # do not perform function calls in argument defaults
    "PLR0913", # too many arguments to function call
    "PLR0912", # too many branches
    "PLR0915", # too many statements
    "PLR2004", # magic value used in comparison
    "FBT001",  # boolean positional arg in function definition
    "FBT002",  # boolean default positional argument in function definition
    "ARG001",  # unused function argument (FastAPI handlers often have unused args)
]
unfixable = [
    "F841", # remove unused variables
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"src/tests/**/*" = ["PLR2004", "S101", "TID252"]
"alembic/**/*" = ["E402"]  # Allow imports after path manipulation in alembic

# MyPy configuration
[tool.mypy]
python_version = "3.13"
strict = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_return_any = true
warn_unreachable = true
show_error_codes = true
show_column_numbers = true
show_error_context = true
pretty = true

# SQLModel specific configuration
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "celery.*",
    "redis.*",
    "alembic.*",
]
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--disable-warnings",
    "-ra",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
testpaths = ["src/tests"]
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

# Coverage configuration
[tool.coverage.run]
source = ["src"]
branch = true
omit = [
    "src/tests/*",
    "src/*/migrations/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
show_missing = true
skip_covered = false
sort = "Cover"

[tool.coverage.html]
directory = "htmlcov"

# Bandit security configuration
[tool.bandit]
exclude_dirs = ["src/tests"]
skips = ["B101", "B601"]  # Skip assert_used and shell_injection_subprocess (dev/test contexts)
